% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_plot_control.R
\name{cnv_plot_control}
\alias{cnv_plot_control}
\title{cnv_plot_control}
\usage{
cnv_plot_control(X_healthy_100KCNV)
}
\arguments{
\item{X_healthy_100KCNV}{matrix to plot cnv}
}
\value{
he_plot
}
\description{
NO4.STEP
}
