% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_plot_test.R
\name{cnv_plot_test}
\alias{cnv_plot_test}
\title{cnv_plot_test}
\usage{
cnv_plot_test(cnv_test_matrix, he_raw_out, he_zero_row)
}
\arguments{
\item{cnv_test_matrix}{matrix to plot cnv}

\item{he_raw_out}{to store the control median}

\item{he_zero_row}{to store the rows zero exists for control&&test groups (should be a list)}
}
\value{
he_plot
}
\description{
NO4.STEP
}
