% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataframe}
\alias{dataframe}
\title{dataframe}
\format{
An object of class \code{data.frame} with 32 rows and 11 columns.
}
\usage{
dataframe
}
\description{
Just test dataframe(Description)
}
\examples{
  head(dataframe)
}
\keyword{datasets}
