% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frag_plot.R
\name{frag_plot}
\alias{frag_plot}
\title{frag_plot}
\usage{
frag_plot(frag_matrix, N_health_samples, N_cancer_samples, type)
}
\arguments{
\item{frag_matrix}{the all matrix for the plot including control && test samples}

\item{N_health_samples}{number of samples of control type}

\item{N_cancer_samples}{number of samples of health type}

\item{type}{type of samples selected 'control' or 'test' (for the color of plot)}
}
\value{
df_ratio
}
\description{
NO3.STEP
}
