% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frag_ratio.R
\name{frag_ratio}
\alias{frag_ratio}
\title{tss_top_genes}
\usage{
frag_ratio(orig_file_path, N_health_samples, N_cancer_samples)
}
\arguments{
\item{orig_file_path}{original file path for all frag.csv files}

\item{N_health_samples}{number of health samples}

\item{N_cancer_samples}{number of cancer samples}
}
\value{
df_ratio
}
\description{
NO2.STEP
}
