% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_top_genes.R
\name{tss_top_genes}
\alias{tss_top_genes}
\title{tss_top_genes}
\usage{
tss_top_genes(test, control, ntop_gene)
}
\arguments{
\item{test}{patient matrix}

\item{control}{healthy matrix}

\item{ntop_gene}{the gene numbers used in the heatmap}
}
\value{
table_cal_all
}
\description{
NO1.STEP
}
